Move pin service to km ownership
