+++ apps/desktop/src/key-management/lock/services/desktop-lock-component.service.ts
@@ -1,11 +1,9 @@
 import { inject } from "@angular/core";
 import { combineLatest, defer, map, Observable } from "rxjs";
 
-import {
-  PinServiceAbstraction,
-  UserDecryptionOptionsServiceAbstraction,
-} from "@bitwarden/auth/common";
+import { UserDecryptionOptionsServiceAbstraction } from "@bitwarden/auth/common";
 import { DeviceType } from "@bitwarden/common/enums";
+import { PinServiceAbstraction } from "@bitwarden/common/key-management/pin/pin.service.abstraction";
 import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/platform-utils.service";
 import { UserId } from "@bitwarden/common/types/guid";
 import { BiometricsService, BiometricsStatus } from "@bitwarden/key-management";
