+++ libs/auth/src/common/abstractions/index.ts
@@ -1,5 +1,4 @@
 export * from "./auth-request-api.service";
-export * from "./pin.service.abstraction";
 export * from "./login-email.service";
 export * from "./login-strategy.service";
 export * from "./user-decryption-options.service.abstraction";
