+++ libs/auth/src/common/services/index.ts
@@ -1,4 +1,3 @@
-export * from "./pin/pin.service.implementation";
 export * from "./login-email/login-email.service";
 export * from "./login-strategies/login-strategy.service";
 export * from "./user-decryption-options/user-decryption-options.service";
