+++ libs/common/src/auth/services/user-verification/user-verification.service.ts
@@ -15,9 +15,9 @@ import {
 } from "@bitwarden/key-management";
 
 // FIXME: remove `src` and fix import
-// eslint-disable-next-line no-restricted-imports
-import { PinServiceAbstraction } from "../../../../../auth/src/common/abstractions/pin.service.abstraction";
+ 
 import { InternalMasterPasswordServiceAbstraction } from "../../../key-management/master-password/abstractions/master-password.service.abstraction";
+import { PinServiceAbstraction } from "../../../key-management/pin/pin.service.abstraction";
 import { I18nService } from "../../../platform/abstractions/i18n.service";
 import { HashPurpose } from "../../../platform/enums";
 import { UserId } from "../../../types/guid";
